modification to the login frontend

+++ task_manager/tasks/views.py
@@ -4,23 +4,10 @@ from django.contrib.auth.forms import UserCreationForm
 from django.shortcuts import render, get_object_or_404, redirect
 from .models import Task
 from .forms import TaskForm
-from django.contrib.auth.decorators import login_required, user_passes_test
+from django.contrib.auth.decorators import login_required
+from django.utils import timezone
 
 
-def is_admin(user):
-    return user.groups.filter(name='Admin').exists()
-
-def is_user(user):
-    return user.groups.filter(name='Usuario').exists()
-
-def is_consultant(user):
-    return user.groups.filter(name='Consultor').exists()
-
-def is_admin_or_user(user):
-    return is_admin(user) or is_user(user)
-
-def is_admin_or_user_or_consultant(user):
-    return is_admin(user) or is_user(user) or is_consultant(user)
 
 # Vista para listar todas las tareas
 @login_required
@@ -30,55 +17,93 @@ def task_list(request):
     print(permisos)
     tasks = Task.objects.all()
     for task in tasks:
-        duration = task.duration()  # Obtener la duración en timedelta
-        if duration:
+         duration = task.duration  # Acceder como propiedad, sin paréntesis
+         if duration:
             # Calcular días, horas y minutos
-            task.duration_days = duration.days
-            task.duration_hours = duration.seconds // 3600
-            task.duration_minutes = (duration.seconds % 3600) // 60
-        else:
+            task.duration_days = duration // (24 * 3600)  # Días
+            task.duration_hours = (duration % (24 * 3600)) // 3600  # Horas
+            task.duration_minutes = (duration % 3600) // 60  # Minutos
+         else:
             task.duration_days = task.duration_hours = task.duration_minutes = None
 
     return render(request, 'tasks/task_list.html', {'tasks': tasks, "permisos": permisos})
 
+
 # Vista para crear una nueva tarea
 @login_required
-@user_passes_test(is_admin_or_user)
 def task_create(request):
+    user = request.user
+    groups = user.groups.all()
+    grupos = [grupo.name for grupo in groups]
+    print("Grupos del usuario:", grupos)
+
+    # Redirigir a la lista de tareas si el usuario es un consultor
+    if "Consultor" in grupos:
+        return redirect("task_list")
+
     if request.method == 'POST':
         form = TaskForm(request.POST)
         if form.is_valid():
-            form.save()
+            task = form.save(commit=False)  # No guardes inmediatamente
+            # Aquí no es necesario asignar created_at, ya que está manejado por el modelo
+            task.save()  # Ahora guarda la tarea
+            messages.success(request, "Tarea creada exitosamente.")  # Mensaje de éxito
             return redirect('task_list')
+        else:
+            messages.error(request, "Error al crear la tarea. Por favor verifica los datos.")  # Mensaje de error
     else:
         form = TaskForm()
+
     return render(request, 'tasks/task_form.html', {'form': form})
+
+
 
 # Vista para editar una tarea existente
 @login_required
-@user_passes_test(is_admin_or_user)
 def task_edit(request, pk):
     user = request.user
-    group_permissions = user.get_group_permissions()
-    if 'tasks.change_task' in group_permissions:
-        task = get_object_or_404(Task, pk=pk)
+    print("Usuario username:", user.username)
+    print("Usuario autenticado:", user.is_authenticated)
 
-        if request.method == 'POST':
-            form = TaskForm(request.POST, instance=task)
-            if form.is_valid():
-                form.save()
-                return redirect('task_list')  # Redirige a la lista de tareas después de guardar.
-        else:
-            form = TaskForm(instance=task)
+    if not  user.is_authenticated:
+        return redirect("login")
 
-        return render(request, 'tasks/task_edit.html', {'form': form})
+    groups = user.groups.all()
+    grupos = [grupo.name for grupo in groups]
+    print("Grupos del usuario:", grupos)
+
+    if "Consultor" in grupos:
+        return redirect("task_list")
+
+    task = get_object_or_404(Task, pk=pk)
+
+    if request.method == 'POST':
+        form = TaskForm(request.POST, instance=task)
+        if form.is_valid():
+            form.save()
+            return redirect('task_list')  # Redirige a la lista de tareas después de guardar.
     else:
-        return redirect('task_list')
+        form = TaskForm(instance=task)
+
+    return render(request, 'tasks/task_edit.html', {'form': form})
+
+
+
 
 # Vista para eliminar una tarea
 @login_required
-@user_passes_test(is_admin)
 def task_delete(request, pk):
+    user = request.user
+    groups = user.groups.all()
+    grupos = [grupo.name for grupo in groups]
+    print("Grupos del usuario:", grupos)
+
+    if "Consultor" in grupos:
+        return redirect("task_list")
+
+    if "Usuario" in grupos:
+        return redirect("task_list")
+
     task = get_object_or_404(Task, pk=pk)
     if request.method == 'POST':
         task.delete()
@@ -88,12 +113,10 @@ def task_delete(request, pk):
 
 # Vista para ver los detalles de una tarea específica
 @login_required
-@user_passes_test(is_admin_or_user_or_consultant)
 def task_detail(request, pk):
     permisos = request.user.get_all_permissions()
     task = get_object_or_404(Task, pk=pk)
     return render(request, 'tasks/task_detail.html', {'task': task,'permisos':permisos})
-
